+++ elastic4s-tcp/src/main/scala/com/sksamuel/elastic4s/searches/queries/matches/MultiMatchQueryBuilderFn.scala
@@ -10,8 +10,8 @@ object MultiMatchQueryBuilderFn {
     q.fuzzyRewrite.foreach(_builder.fuzzyRewrite)
     q.analyzer.foreach(_builder.analyzer)
     q.fields.foreach {
-      case (name, -1F) => _builder.field(name)
-      case (name, boost) => _builder.field(name, boost)
+      case (name, -1D) => _builder.field(name)
+      case (name, boost) => _builder.field(name, boost.toFloat)
     }
     q.cutoffFrequency.map(_.toFloat).foreach(_builder.cutoffFrequency)
     q.fuzziness.foreach(_builder.fuzziness)
